+++ apps/client/src/components/option/index.tsx
@@ -1,5 +1,5 @@
 "use client"
-import { CodeSandboxLogoIcon } from "@radix-ui/react-icons"
+import { CheckCircledIcon, CircleIcon, CodeSandboxLogoIcon } from "@radix-ui/react-icons"
 import { Button, Col } from "react-bootstrap"
 
 interface OptionProps {
@@ -12,8 +12,8 @@ interface OptionProps {
 export default function Option({ isSelected, value, label, handleChange }: OptionProps) {
   return (
     <Col>
-      <Button variant={isSelected ? "primary" : "secondary"} className="btn-block" onClick={() => handleChange(value)}>
-        <CodeSandboxLogoIcon className="icon-left" />{label}
+      <Button variant="secondary" className="btn-block" onClick={() => handleChange(value)}>
+        {isSelected ? <CheckCircledIcon className="icon-left" /> : <CircleIcon className="icon-left" />}{label}
       </Button>
     </Col>
   )